requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
psycopg2-binary==2.9.7
pandas==2.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
Werkzeug==2.3.7
reportlab==4.0.4
xlsxwriter==3.1.3
xlrd==2.0.1
gunicorn==21.2.0